requirements.txt
PyYAML==6.0
requests==2.27.1
python-environ==0.4.54
coloredlogs==15.0.1
Unidecode==1.2.0
beautifulsoup4==4.10.0
peewee==3.14.9
selenium==4.1.0
psycopg2-binary==2.9.3
geckodriver-autoinstaller==0.1.0
phonenumbers==8.12.44